Move main2 to main
