+++ lib/main.js
@@ -2,84 +2,53 @@ const fs = require('fs')
 const path = require('path')
 const os = require('os')
 
-function log (message) {
-  console.log(`[dotenv][DEBUG] ${message}`)
-}
+const LINE = /(?:^|^)\s*(?:export\s+)?([\w.-]+)(?:\s*=\s*?|:\s+?)(\s*'(?:\\'|[^'])*'|\s*"(?:\\"|[^"])*"|[^#\r\n]+)?\s*(?:#.*)?(?:$|$)/mg
 
-const NEWLINE = '\n'
-const RE_INI_KEY_VAL = /^\s*([\w.-]+)\s*=\s*("[^"]*"|'[^']*'|.*?)(\s+#.*)?$/
-const RE_NEWLINES = /\\n/g
-const NEWLINES_MATCH = /\r\n|\n|\r/
-
-// Parses src into an Object
-function parse (src, options) {
-  const debug = Boolean(options && options.debug)
-  const multiline = Boolean(options && options.multiline)
+// Parser src into an Object
+function parse (src) {
   const obj = {}
 
-  // convert Buffers before splitting into lines and processing
-  const lines = src.toString().split(NEWLINES_MATCH)
+  // Convert buffer to string
+  let lines = src.toString()
 
-  for (let idx = 0; idx < lines.length; idx++) {
-    let line = lines[idx]
+  // Convert line breaks to same format
+  lines = lines.replace(/\r\n?/mg, '\n')
 
-    // matching "KEY' and 'VAL' in 'KEY=VAL'
-    const keyValueArr = line.match(RE_INI_KEY_VAL)
-    // matched?
-    if (keyValueArr != null) {
-      const key = keyValueArr[1]
-      // default undefined or missing values to empty string
-      let val = (keyValueArr[2] || '')
-      let end = val.length - 1
-      const isDoubleQuoted = val[0] === '"' && val[end] === '"'
-      const isSingleQuoted = val[0] === "'" && val[end] === "'"
+  let match
+  while ((match = LINE.exec(lines)) != null) {
+    const key = match[1]
 
-      const isMultilineDoubleQuoted = val[0] === '"' && val[end] !== '"'
-      const isMultilineSingleQuoted = val[0] === "'" && val[end] !== "'"
+    // Default undefined or null to empty string
+    let value = (match[2] || '')
 
-      // if parsing line breaks and the value starts with a quote
-      if (multiline && (isMultilineDoubleQuoted || isMultilineSingleQuoted)) {
-        const quoteChar = isMultilineDoubleQuoted ? '"' : "'"
+    // Remove whitespace
+    value = value.trim()
 
-        val = val.substring(1)
+    // Check if double quoted
+    const maybeQuote = value[0]
 
-        while (idx++ < lines.length - 1) {
-          line = lines[idx]
-          end = line.length - 1
-          if (line[end] === quoteChar) {
-            val += NEWLINE + line.substring(0, end)
-            break
-          }
-          val += NEWLINE + line
-        }
-      // if single or double quoted, remove quotes
-      } else if (isSingleQuoted || isDoubleQuoted) {
-        val = val.substring(1, end)
+    // Remove surrounding quotes
+    // value = value.replace(/^(['"])(.*)\1$/mg, '$2')
+    value = value.replace(/^(['"])([\s\S]+)\1$/mg, '$2')
 
-        // if double quoted, expand newlines
-        if (isDoubleQuoted) {
-          val = val.replace(RE_NEWLINES, NEWLINE)
-        }
-      } else {
-        // remove surrounding whitespace
-        val = val.trim()
-      }
-
-      obj[key] = val
-    } else if (debug) {
-      const trimmedLine = line.trim()
-
-      // ignore empty and commented lines
-      if (trimmedLine.length && trimmedLine[0] !== '#') {
-        log(`Failed to match key and value when parsing line ${idx + 1}: ${line}`)
-      }
+    // Expand newlines if double quoted
+    if (maybeQuote === '"') {
+      value = value.replace(/\\n/g, '\n')
+      value = value.replace(/\\r/g, '\r')
     }
+
+    // Add to object
+    obj[key] = value
   }
 
   return obj
 }
 
-function resolveHome (envPath) {
+function _log (message) {
+  console.log(`[dotenv][DEBUG] ${message}`)
+}
+
+function _resolveHome (envPath) {
   return envPath[0] === '~' ? path.join(os.homedir(), envPath.slice(1)) : envPath
 }
 
@@ -89,11 +58,10 @@ function config (options) {
   let encoding = 'utf8'
   const debug = Boolean(options && options.debug)
   const override = Boolean(options && options.override)
-  const multiline = Boolean(options && options.multiline)
 
   if (options) {
     if (options.path != null) {
-      dotenvPath = resolveHome(options.path)
+      dotenvPath = _resolveHome(options.path)
     }
     if (options.encoding != null) {
       encoding = options.encoding
@@ -101,8 +69,8 @@ function config (options) {
   }
 
   try {
-    // specifying an encoding returns a string instead of a buffer
-    const parsed = DotenvModule.parse(fs.readFileSync(dotenvPath, { encoding }), { debug, multiline })
+    // Specifying an encoding returns a string instead of a buffer
+    const parsed = DotenvModule.parse(fs.readFileSync(dotenvPath, { encoding }))
 
     Object.keys(parsed).forEach(function (key) {
       if (!Object.prototype.hasOwnProperty.call(process.env, key)) {
@@ -114,9 +82,9 @@ function config (options) {
 
         if (debug) {
           if (override === true) {
-            log(`"${key}" is already defined in \`process.env\` and WAS overwritten`)
+            _log(`"${key}" is already defined in \`process.env\` and WAS overwritten`)
           } else {
-            log(`"${key}" is already defined in \`process.env\` and was NOT overwritten`)
+            _log(`"${key}" is already defined in \`process.env\` and was NOT overwritten`)
           }
         }
       }
@@ -125,7 +93,7 @@ function config (options) {
     return { parsed }
   } catch (e) {
     if (debug) {
-      log(`Failed to load ${dotenvPath} ${e.message}`)
+      _log(`Failed to load ${dotenvPath} ${e.message}`)
     }
 
     return { error: e }
