+++ tests/.env
@@ -16,14 +16,11 @@ DONT_EXPAND_SQUOTED='dontexpand\nnewlines'
 INLINE_COMMENTS=inline comments # work #very #well
 INLINE_COMMENTS_SINGLE_QUOTES='inline comments outside of #singlequotes' # work
 INLINE_COMMENTS_DOUBLE_QUOTES="inline comments outside of #doublequotes" # work
-INLINE_COMMENTS_SPACE=inline comments must start with#space
+INLINE_COMMENTS_SPACE=inline comments dont have to start with a#space
 EQUAL_SIGNS=equals==
 RETAIN_INNER_QUOTES={"foo": "bar"}
-RETAIN_LEADING_DQUOTE="retained
-RETAIN_LEADING_SQUOTE='retained
-RETAIN_TRAILING_DQUOTE=retained"
-RETAIN_TRAILING_SQUOTE=retained'
 RETAIN_INNER_QUOTES_AS_STRING='{"foo": "bar"}'
 TRIM_SPACE_FROM_UNQUOTED=    some spaced out string
 USERNAME=therealnerdybeast@example.tld
     SPACED_KEY = parsed
+UNQUOTED_ESCAPE_CHARACTERS=bar\\ bar
