+++ tests/.env-multiline
@@ -30,7 +30,3 @@ IS
 A
 MULTILINE
 STRING'
-
-MULTI_UNENDED="THIS
-LINE HAS
-NO END QUOTE
