+++ tests/.env-multiline.backup
@@ -30,3 +30,7 @@ IS
 A
 MULTILINE
 STRING'
+
+MULTI_UNENDED="THIS
+LINE HAS
+NO END QUOTE
