+++ tests/test-config.js.backup
@@ -5,11 +5,13 @@ const path = require('path')
 const sinon = require('sinon')
 const t = require('tap')
 
-const dotenv = require('../lib/main2')
+const dotenv = require('../lib/main')
 
 const mockParseResponse = { test: 'foo' }
 let readFileSyncStub
 let parseStub
+
+t.plan(14)
 
 t.beforeEach(() => {
   readFileSyncStub = sinon.stub(fs, 'readFileSync').returns('test=foo')
@@ -59,6 +61,13 @@ t.test('takes option for debug', ct => {
 
   ct.ok(logStub.called)
   logStub.restore()
+})
+
+t.test('takes option for multiline', ct => {
+  ct.plan(1)
+  const testMultiline = true
+  dotenv.config({ multiline: testMultiline })
+  ct.equal(parseStub.args[0][1].multiline, testMultiline)
 })
 
 t.test('reads path with encoding, parsing output to process.env', ct => {
